Made a few changes to update the code, consistent with the most recent anura changes.

+++ src/kre/RenderTarget.cpp
@@ -21,39 +21,39 @@
 	   distribution.
 */
 
-#include "../asserts.hpp"
+#include "asserts.hpp"
 #include "DisplayDevice.hpp"
 #include "RenderTarget.hpp"
-#include "../variant_utils.hpp"
+#include "variant_utils.hpp"
 
 namespace KRE
 {
-	RenderTarget::RenderTarget(unsigned width, unsigned height, 
+	RenderTarget::RenderTarget(int width, int height, 
 		unsigned color_plane_count, 
 		bool depth, 
 		bool stencil, 
 		bool use_multi_sampling, 
 		unsigned multi_samples)
 		: width_(width),
-		height_(height),
-		color_attachments_(color_plane_count),
-		depth_attachment_(depth),
-		stencil_attachment_(stencil),
-		multi_sampling_(use_multi_sampling),
-		multi_samples_(multi_samples),
-		clear_color_(0.0f, 0.0f, 0.0f, 1.0f)
+		  height_(height),
+		  color_attachments_(color_plane_count),
+		  depth_attachment_(depth),
+		  stencil_attachment_(stencil),
+		  multi_sampling_(use_multi_sampling),
+		  multi_samples_(multi_samples),
+		  clear_color_(0.0f, 0.0f, 0.0f, 1.0f)
 	{
 	}
 
 	RenderTarget::RenderTarget(const variant& node)
 		: width_(0),
-		height_(0),
-		color_attachments_(1),
-		depth_attachment_(false),
-		stencil_attachment_(false),
-		multi_sampling_(false),
-		multi_samples_(0),
-		clear_color_(0.0f, 0.0f, 0.0f, 1.0f)
+		  height_(0),
+		  color_attachments_(1),
+		  depth_attachment_(false),
+		  stencil_attachment_(false),
+		  multi_sampling_(false),
+		  multi_samples_(0),
+		  clear_color_(0.0f, 0.0f, 0.0f, 1.0f)
 	{
 		ASSERT_LOG(node.is_map(), "RenderTarget definitions must be maps: " << node.to_debug_string());
 		ASSERT_LOG(node.has_key("width"), "Render target must have a 'width' attribute.");
@@ -126,10 +126,10 @@ namespace KRE
 	variant RenderTarget::write()
 	{
 		variant_builder res;
-		res.add("width", static_cast<int>(width_));
-		res.add("height", static_cast<int>(height_));
+		res.add("width", width_);
+		res.add("height", height_);
 		if(color_attachments_ != 1) {
-			res.add("color_planes", static_cast<int>(color_attachments_));
+			res.add("color_planes", color_attachments_);
 		}
 		if(depth_attachment_) {
 			res.add("depth_buffer", variant::from_bool(depth_attachment_));
@@ -139,7 +139,7 @@ namespace KRE
 		}
 		if(multi_sampling_) {
 			res.add("use_multisampling", variant::from_bool(multi_sampling_));
-			res.add("samples", static_cast<int>(multi_samples_));
+			res.add("samples", multi_samples_);
 		}
 		return res.build();
 	}
